requirements.txt
sentence-transformers==3.3.1
azure-identity==1.19.0
pandas==2.2.3
requests==2.32.3
openai==1.59.6
